requirements.txt
arrow==0.14.7; python_version < '3.0'
arrow==1.2.3; python_version >= '3.0' and python_version <= '3.7'
arrow==1.3.0; python_version >= '3.8'
bandit==1.6.2; python_version < '3.0'
bandit==1.7.1; python_version == '3.6'
bandit==1.7.5; python_version == '3.7'
bandit==1.7.10; python_version == '3.8'
bandit==1.8.6; python_version >= '3.9'
Django==1.11.29; python_version < '3.0'
Django==3.2.25; python_version >= '3.0' and python_version <= '3.7'
Django==4.2.24; python_version >= '3.8' and python_version <= '3.9'
Django==5.2.6; python_version >= '3.10'
iso8601==2.1.0; python_version >= '3.7'
django-prettyjson==0.4.1
phonenumbers==9.0.14
psycopg2-binary==2.8.6; python_version < '3.0'
psycopg2-binary==2.9.8; python_version >= '3.0' and python_version < '3.7'
psycopg2-binary==2.9.9; python_version == '3.7'
psycopg2-binary==2.9.10; python_version >= '3.8'
pylint==3.3.8; python_version >= '3.9'
pylint==3.2.7; python_version == '3.8'
pylint==2.17.7; python_version == '3.7'
pylint==2.13.9; python_version == '3.6'
pylint==1.9.5; python_version < '3.0'
